+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - (Feature) AgencyCache Interface
 - (Feature) Agency Cache Poll EE Extension
 - (Feature) Metrics Counter
+- (Feature) Requests Bytes Counter
 
 ## [1.2.29](https://github.com/arangodb/kube-arangodb/tree/1.2.29) (2023-06-08)
 - (Maintenance) Add govulncheck to pipeline, update golangci-linter